refactor: use a controller model for wizard flows
This patch uses a Controller pattern for managing the flow of creating a
new bucket. The value is that the TreeView nodes aren't storing
unrelated logic for other extension functionality. This pattern, applied
across all of these flows, should organize the code in a more
maintainable way.
